requirements.txt
flask>=3.0.0
nltk>=3.8.1
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
Pillow>=10.0.0

# LLM providers — install only what you need
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
